Updated blocks (and supporting files) to reference aem.js instead of lib-franklin.js so copy/paste of blocks from the collection work on new projects from boilerplate moving forward
