+++ head.html
@@ -1,5 +1,5 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/scripts/lib-franklin.js" type="module" crossorigin="use-credentials"></script>
+<script src="/scripts/aem.js" type="module" crossorigin="use-credentials"></script>
 <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
 <link rel="stylesheet" href="/styles/styles.css"/>
 <link rel="icon" href="data:,">
