+++ test/blocks/footer/footer.test.js
@@ -6,7 +6,7 @@ import { expect } from '@esm-bundle/chai';
 
 document.body.innerHTML = await readFile({ path: '../../scripts/dummy.html' });
 
-const { buildBlock, decorateBlock, loadBlock } = await import('../../../scripts/lib-franklin.js');
+const { buildBlock, decorateBlock, loadBlock } = await import('../../../scripts/aem.js');
 
 document.body.innerHTML = await readFile({ path: '../../scripts/body.html' });
